fix errors in formula testing

+++ src/main/php/model/formula/formula_link_list.php
@@ -137,20 +137,22 @@ class formula_link_list
         global $db_con;
         $result = true;
 
-        foreach ($this->lst as $frm_lnk) {
-            if ($result) {
-                if ($frm_lnk->can_change() > 0 and $frm_lnk->not_used()) {
-                    //$db_con = new mysql;
-                    $db_con->usr_id = $this->usr->id;
-                    // delete first all user configuration that have also been excluded
-                    $db_con->set_type(DB_TYPE_USER_PREFIX . DB_TYPE_FORMULA_LINK);
-                    $result = $db_con->delete(array('formula_link_id', 'excluded'), array($frm_lnk->id, '1'));
-                    if ($result) {
-                        $db_con->set_type(DB_TYPE_FORMULA_LINK);
-                        $result = $db_con->delete('formula_link_id', $frm_lnk->id);
+        if ($this->lst != null) {
+            foreach ($this->lst as $frm_lnk) {
+                if ($result) {
+                    if ($frm_lnk->can_change() > 0 and $frm_lnk->not_used()) {
+                        //$db_con = new mysql;
+                        $db_con->usr_id = $this->usr->id;
+                        // delete first all user configuration that have also been excluded
+                        $db_con->set_type(DB_TYPE_USER_PREFIX . DB_TYPE_FORMULA_LINK);
+                        $result = $db_con->delete(array('formula_link_id', 'excluded'), array($frm_lnk->id, '1'));
+                        if ($result) {
+                            $db_con->set_type(DB_TYPE_FORMULA_LINK);
+                            $result = $db_con->delete('formula_link_id', $frm_lnk->id);
+                        }
+                    } else {
+                        log_err("Cannot delete a formula word link (id " . $frm_lnk->id . "), which is used or created by another user.", "formula_link_list->del_without_log");
                     }
-                } else {
-                    log_err("Cannot delete a formula word link (id " . $frm_lnk->id . "), which is used or created by another user.", "formula_link_list->del_without_log");
                 }
             }
         }
